app/requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
numpy==1.21.2
pandas==1.3.2
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
